esdm_es: fix warning in testing tool

+++ addon/test/getrawentropy.c
@@ -88,7 +88,7 @@ static int getrawentropy(struct opts *opts)
 
 		requested -= (uint32_t)ret;
 
-		fprintf(stderr, "Fetched %zd events, still %lu to do\n", ret / sizeof(uint32_t), requested  / sizeof(uint32_t));
+		fprintf(stderr, "Fetched %lu events, still %lu to do\n", (uint32_t)ret / sizeof(uint32_t), requested  / sizeof(uint32_t));
 	}
 
 	ret = 0;
